requirements.txt
runpod
easyocr
pymupdf
requests
Pillow
